Refactor project name
