+++ StreamNode/Controllers/ApplicationMixer.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="RemoteControl.ApplicationMixer"
+﻿<UserControl x:Class="StreamNode.ApplicationMixer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:RemoteControl"
+             xmlns:local="clr-namespace:StreamNode"
              mc:Ignorable="d">
     <Grid RenderTransformOrigin="0.5,0.5">
         <Grid.RenderTransform>
