+++ StreamNode/Controllers/SceneController.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="RemoteControl.Controllers.SceneController"
+﻿<UserControl x:Class="StreamNode.Controllers.SceneController"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:RemoteControl.Controllers"
+             xmlns:local="clr-namespace:StreamNode.Controllers"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <Grid>
