+++ StreamNode/StreamNode.csproj
@@ -5,11 +5,6 @@
     <UseWPF>true</UseWPF>
     <Authors>Luca Taddeo</Authors>
     <RepositoryUrl>https://github.com/lucalas/StreamNode</RepositoryUrl>
-    <PublishSingleFile>true</PublishSingleFile>
-    <SelfContained>true</SelfContained>
-    <RuntimeIdentifier>win-x64</RuntimeIdentifier>
-    <DebugType>None</DebugType>
-    <DebugSymbols>false</DebugSymbols>
     <Version>0.0.1</Version>
     <Company />
   </PropertyGroup>
